options.loadObjects dependent depth (verification advisable)
I tried to get a quick retrieve of the available address books via
dav.createAccount({accountType : 'carddav', loadCollections: true,
loadObjects: false}), but it wasn't "quick"!
After some debugging I recognized that I also retrieved a list of all
contacts (some hundreds), beside the address books.
I'm not very familiar with WebDAV and also not an RFC extraterrestrial,
but after some reading I interpreted the specs in that way that I only
have to expect the contacts when loadObject = true.
This is what this fix does and now the address book query is quick.